recovery: adaptive reordering thresholds
The idea is to increase packet and time reordering thresholds when
spurious losses are detected (i.e. when a packet previously declared
lost is acked).

In addition, the congestion control state is rolled back to its state
before the last congestion event when a spurious loss is detected.

This is similar to what Chrome currently implements.